Rudimentary chest interaction (+Signs)
Using Bot.StoreItem() will store it's selected item into the Chest in front of it.
Using Bot.TakeItem() will store it's selected item into the Chest in front of it.

Using Bot.PlaceItem() with a Sign ahead will configure the Sign with the selected item as if the player had.